content_recommendation/requirements.txt
# Core dependencies for the recommendation demo app
flet
google-cloud-bigquery
pyyaml
python-dotenv